Move local-storage modules to clearer location
