+++ client/lib/local-storage-bypass/test/index.js
@@ -9,7 +9,7 @@ import { spy } from 'sinon';
 /**
  * Internal dependencies
  */
-import { key, clear, setItem, getItem, removeItem, length } from '../localstorage-bypass';
+import { key, clear, setItem, getItem, removeItem, length } from '..';
 
 describe( 'localstorage-bypass', () => {
 	// Spy on the original localStorage functions
