+++ client/lib/local-storage-polyfill/README.md
@@ -1,7 +1,7 @@
-local-storage
-=============
+local-storage-polyfill
+======================
 
-This is a very simple polyfill that implements the core methods of localStorage when the native API is either unavailable or not working correctly (e.g., safari in private mode):
+This is a very simple polyfill that implements the core methods of localStorage when the native API is either unavailable or not working correctly (e.g., safari in private mode).
 
 ```js
 localStorage.setItem( key, value )
@@ -14,5 +14,5 @@ localStorage.length
 Just require and call this once into the page (prior to any calls to these methods) and it will either create or augment the `window.localStorage` object as necessary.
 
 ```es6
-import localStorePolyfill from 'lib/local-storage';
+import localStorePolyfill from 'lib/local-storage-polyfill';
 ```
