+++ client/lib/local-storage-polyfill/test/index.js
